# perf: [Geneva Exporter] optimize schema creation in OTLP encoder to create only when required