--- a/opentelemetry-exporter-geneva/geneva-uploader/src/bench.rs
+++ b/opentelemetry-exporter-geneva/geneva-uploader/src/bench.rs
@@ -241,4 +241,98 @@ mod benchmarks {
 
         criterion.final_summary();
     }
+
+    /// Benchmark specifically to test schema creation optimization
+    /// This test focuses on scenarios where many log records share the same schema structure,
+    /// which is where our "create schema only when needed" optimization should show the most benefit.
+    #[test]
+    #[ignore = "benchmark on crate private, ignored by default during normal test runs"]
+    // To run: $cargo test --release schema_optimization_benchmark -- --nocapture --ignored
+    fn schema_optimization_benchmark() {
+        let mut criterion = Criterion::default();
+        let encoder = OtlpEncoder::new();
+        let metadata = "namespace=benchmark/eventVersion=Ver1v0";
+
+        // Benchmark: Many logs with identical schema structure (should benefit most from optimization)
+        let mut group = criterion.benchmark_group("schema_optimization");
+        for batch_size in [10, 50, 100, 500, 1000].iter() {
+            group.bench_with_input(
+                BenchmarkId::new("identical_schemas", batch_size),
+                batch_size,
+                |b, &batch_size| {
+                    // Create logs with identical schema structure - all have same 4 attributes
+                    let logs: Vec<LogRecord> = (0..batch_size)
+                        .map(|i| {
+                            let mut log = LogRecord {
+                                observed_time_unix_nano: 1_700_000_000_000_000_000 + i as u64,
+                                severity_number: 9,
+                                severity_text: "INFO".to_string(),
+                                event_name: "IdenticalSchemaEvent".to_string(),
+                                body: Some(AnyValue {
+                                    value: Some(
+                                        opentelemetry_proto::tonic::common::v1::any_value::Value::StringValue(
+                                            "Log message".to_string(),
+                                        ),
+                                    ),
+                                }),
+                                ..Default::default()
+                            };
+
+                            // Add exactly the same 4 attributes to each log to ensure identical schemas
+                            log.attributes.push(KeyValue {
+                                key: "service_name".to_string(),
+                                value: Some(AnyValue {
+                                    value: Some(
+                                        opentelemetry_proto::tonic::common::v1::any_value::Value::StringValue(
+                                            "test-service".to_string(),
+                                        ),
+                                    ),
+                                }),
+                            });
+                            log.attributes.push(KeyValue {
+                                key: "request_id".to_string(),
+                                value: Some(AnyValue {
+                                    value: Some(
+                                        opentelemetry_proto::tonic::common::v1::any_value::Value::StringValue(
+                                            format!("req_{}", i),
+                                        ),
+                                    ),
+                                }),
+                            });
+                            log.attributes.push(KeyValue {
+                                key: "status_code".to_string(),
+                                value: Some(AnyValue {
+                                    value: Some(
+                                        opentelemetry_proto::tonic::common::v1::any_value::Value::IntValue(200),
+                                    ),
+                                }),
+                            });
+                            log.attributes.push(KeyValue {
+                                key: "duration_ms".to_string(),
+                                value: Some(AnyValue {
+                                    value: Some(
+                                        opentelemetry_proto::tonic::common::v1::any_value::Value::DoubleValue(
+                                            (i % 100) as f64 + 50.5,
+                                        ),
+                                    ),
+                                }),
+                            });
+
+                            log
+                        })
+                        .collect();
+
+                    b.iter(|| {
+                        let res = black_box(
+                            encoder.encode_log_batch(black_box(logs.iter()), black_box(metadata)),
+                        );
+                        black_box(res);
+                    });
+                },
+            );
+        }
+        group.finish();
+
+        criterion.final_summary();
+    }
 }