## Changes

Optimized the OTLP encoder to create schema entries only when they don't already exist in the batch, eliminating unnecessary schema creation for duplicate schemas.

### Changes

- __Modified `otlp_encoder.rs`__: Moved `Self::create_schema()` call inside the conditional check that verifies if a schema already exists
- __Added comprehensive benchmark__: New `schema_optimization_benchmark` to measure performance impact on identical schema scenarios
- __Performance improvement__: 69-74% faster processing for batches with identical schemas


### Performance Impact

Benchmark results for logs with identical schema structures:

| Batch Size | Before | After | Improvement |
|------------|--------|-------|-------------|
| 10 logs | 51.8 µs | 16.1 µs | **69% faster** |
| 50 logs | 261.7 µs | 69.5 µs | **73% faster** |
| 100 logs | 527.8 µs | 138.4 µs | **74% faster** |
| 500 logs | 2.64 ms | 679.7 µs | **74% faster** |
| 1000 logs | 5.14 ms | 1.33 ms | **74% faster** |

## Merge requirement checklist

* [ ] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust-contrib/blob/main/CONTRIBUTING.md) guidelines followed
* [ ] Unit tests added/updated (if applicable)
* [ ] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [ ] Changes in public API reviewed (if applicable)